Project Files
This commit introduces the initial upload of all the essential project files to the AWLGrowthCalculator repository. These files will lay the groundwork for our application's development, execution, and distribution. The upload includes:

AWLGrowthCalculator.sln: This is the solution file for the project. It holds the configuration and metadata for the entire solution, such as project dependencies, build configurations, and version control settings.

AWLGrowthCalculator.csproj: This is the C# project file, containing detailed information about the project, including target framework, package references, and other project-specific settings.

/bin: This directory is part of the .NET build output folders. It contains the binary files, DLLs, and EXEs, resulting from building the project. Note that these files are often not included in version control as they can be regenerated through build processes.

/obj: This directory also stores temporary object files and other files used to create the final binary during the build process. Similar to the /bin directory, it is common practice to exclude this directory from version control as the files within it are dynamically generated.

Other Essential Files: Several other important project files and directories are also included in this upload, such as class files (.cs), interface files (.cs), and possibly resource files.

This initial commit establishes the foundation for our application and is crucial for other developers to clone the project and begin contributing or testing the application in their local development environments. Please note that the /bin and /obj directories are typically not included in the version control due to their dynamic nature and are thus omitted here, which is standard practice in .NET development.